src/modules/gastrointestinal_disorder_diagnosis_support/requirements.txt
streamlit
fastapi
uvicorn
pymongo
python-dotenv
requests
pydantic